airflow-dags-master/requirements.txt
snowflake-connector-python==2.4.1
snowflake-sqlalchemy==1.2.4

apache-airflow==2.0.2
apache-airflow-providers-amazon==1.4
xgboost==1.3.3
smart_open[s3]==5.1.0
pandas==1.3.0
boto3==1.17.106
numpy==1.20.0
numba==0.54.0
apache-airflow-providers-slack==3.0.0
scikit-learn==0.24.2
shap==0.39.0
tabulate==0.8.9

imblearn==0.0
python-dotenv==0.19.0
matplotlib==3.4.2
seaborn==0.11.1

moto==2.0.11
botocore==1.20.106